backend/requirements/base.txt
# ============================================================
# HealthTech Solutions — Dependencias Base
# Compatibles con Oracle 19c RAC y 21c XE
# ============================================================

# Framework
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3

# Oracle Driver — OBLIGATORIO python-oracledb (no cx_Oracle)
python-oracledb==2.3.0

# Autenticación JWT
djangorestframework-simplejwt==5.3.1

# Variables de entorno
python-decouple==3.8

# Validaciones y utilidades
Pillow==10.4.0          # Manejo de imágenes
phonenumbers==8.13.43   # Validación teléfonos

# AWS S3 (PACS imágenes médicas)
boto3==1.35.0
django-storages==1.14.4

# Reportes
reportlab==4.2.2        # Generación PDF
openpyxl==3.1.5         # Exportación Excel

# Seguridad
cryptography==43.0.0

# Logging y monitoreo
django-auditlog==3.0.0
